requirements.txt
Flask==3.1.1
Flask-Limiter==3.12
qrcode==8.2
pillow==11.3.0
Flask-SocketIO==5.5.1
eventlet==0.39.1
python-dotenv==1.1.0
pyotp==2.9.0
requests==2.32.4
limits[redis]==5.4.0